+++ CHANGELOG.md
@@ -5,3 +5,4 @@
 - [docs:prevent the release of the final version of the sdk.](https://github.com/Tencent/spring-cloud-tencent/pull/944)
 - [feat:support webclient and gateway report call metrics](https://github.com/Tencent/spring-cloud-tencent/pull/946)
 - [fix: fix nacos CircuitBreaker disable bug.](https://github.com/Tencent/spring-cloud-tencent/pull/949)
+- [feature: add config for customized local port.](https://github.com/Tencent/spring-cloud-tencent/pull/957)